pyproject.toml
[project]
name = "llm-testing-framework"
version = "0.1.0"
description = "A modern framework for testing and evaluating LLMs"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic_settings",
    "pandas",
    "numpy",  # Часто нужен вместе с pandas
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.32.0",
    "plotly>=5.18.0",
    "altair>=5.3.0",
    "streamlit-autorefresh>=0.1.0",
]
core = [
    "transformers>=4.38.0",
    "torch>=2.0.0",
]
db = [
    "motor>=3.3.2"
]
celery = [
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
]
adapters = [
    "ollama>=0.1.8",
    "huggingface-hub>=0.20.0",
    "aiohttp",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart"
]
eval = [ # Вынесем зависимости для evaluation в отдельную группу
    "nltk>=3.8.1",
    "rouge-score>=0.1.2",
]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.3.0",
]

all = [
    "llm-testing-framework[ui,core,db,celery,adapters,api,eval]",
]

[tool.setuptools.packages.find]
where = ["src"]